// apps/web/src/lib/generated/models/PipelineLandscapeBucketRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PipelineLandscapeBucketRead = {
  count: number;
  entity_id?: (string | null);
  key: string;
  label: string;
  phase_counts?: Record<string, number>;
  share: number;
};